% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GarciaF200}
\alias{GarciaF200}
\title{GarciaF200  sub data set from Garcia-Franco et al. (2021)}
\format{
A nested list with 3 sub-lists:
\describe{
  \item{1 to 3}{A list containing the vegetation NMR spectrum of one of the following sites.
    \describe{
      \item{name}{"EB_Vegetation", "Fendt_Vegetation", "Graswang_Vegetation"}
      \item{raw.spec}{A data frame with 2 columns:
        \describe{
          \item{ppm}{Numeric vector.}
          \item{raw.intensity}{Numeric vector.}
        }
      }
    }
  }
}
}
\source{
Garcia-Franco et al. (2021) DOI: 10.1007/s00374-020-01518-0
}
\usage{
GarciaF200
}
\description{
Contains 3 CP MAS 13C NMR spectra.
}
\details{
The spectra were taken in a NMR spectrometer with field strength of 200 MHz and MAS rate of 6.8 kHz
}
\examples{
data(GarciaF200)
}
\keyword{datasets}
