% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hall300}
\alias{Hall300}
\title{Hall sub data set from Hall et al. (2020)}
\format{
A nested list with 17 sub-lists:
\describe{
  \item{1 to 17}{A list containing the soil NMR spectrum of one of the following sites.
    \describe{
      \item{name}{"Calhoun", "CPER", "DCFS", "elve", "GRSM", "HARV", "icac", "JERC", "KONZ", "LENO", "MOAB", "NIWO", "ONAQ", "samt", "SCBI", "UNDE", "WOOD"}
      \item{raw.spec}{A data frame with 2 columns:
        \describe{
          \item{ppm}{Numeric vector.}
          \item{raw.intensity}{Numeric vector.}
        }
      }
    }
  }
}
}
\source{
https://portal.edirepository.org/nis/mapbrowse?packageid=edi.575.1
}
\usage{
Hall300
}
\description{
Contains 17 CP MAS 13C NMR spectra.
}
\details{
The spectra were taken in a NMR spectrometer with field strength of 300 MHz and MAS rate of 12 kHz
}
\examples{
data(Hall300)
}
\keyword{datasets}
