% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_nmr.R
\name{int_nmr}
\alias{int_nmr}
\title{Integration function}
\usage{
int_nmr(raw.spec, NMRmeth = NULL, SSBcorr = FALSE)
}
\arguments{
\item{raw.spec}{Raw spectrum}

\item{NMRmeth}{Regions to be integrated.
Default is spinning side bands, other methods available include: Bonanomi ("Bonanomi") and Molecular mixing model ("MMM").}
}
\description{
This function allows you to integrate the 13C-NMR spectra using diferent integration regions.
The loaded Raw spectra can be integrated using the spinning side bands regions(default),
the Bonanomi("Bonanomi") regions or the Molecular Mixing Model regions("MMM").
The function returns the corrected, normalized and flattened spectrum
}
\keyword{integration}
\keyword{normalization,}
