% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMM_fit.R
\name{MMM_fit}
\alias{MMM_fit}
\title{All combination fitting of NMR spectra.}
\usage{
MMM_fit(sample, standards, ex.smaller = NULL, NMRmeth, FixNC)
}
\arguments{
\item{sample}{Sample Integrals}

\item{standards}{List of all standards}

\item{ex.smaller}{Exclude portions smaller than a given value (decimal form), default to NULL}

\item{NMRmeth}{Regions to be integrated.
Default is spinning side bands, other methods available include: Bonanomi ("Bonanomi") and Molecular mixing model ("MMM").}

\item{FixNC}{TRUE or FALSE, for fixing or not the NC ratio on the sample fitting.}
}
\description{
The function wraps the Linear combination fit of the integrated regions of the molecular mixing model.
}
\keyword{correction}
\keyword{normalization}
