% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMM_solve_QP.R
\name{MMM_solve_QP}
\alias{MMM_solve_QP}
\title{Linear combination fitting solve function}
\usage{
MMM_solve_QP(LCF.stds, LCF.samp, NMRmeth = NULL, FixNC)
}
\arguments{
\item{LCF.stds}{Standards for LCF}

\item{LCF.samp}{NMR Sample(s) for LCF}

\item{NMRmeth}{Regions to be integrated.
Default is spinning side bands, other methods available include: Bonanomi ("Bonanomi") and Molecular mixing model ("MMM").}
}
\description{
Quadratic programming solution function for linear combination fitting (LCF)
}
\keyword{correction}
\keyword{normalization,}
