% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_data.R
\name{nc_data}
\alias{nc_data}
\title{N/C data merge function}
\usage{
nc_data(NCdata)
}
\arguments{
\item{NCdata}{Raw spectrum}
}
\description{
This function allows you import a .csv file and create a dataframe with the C and N data of the samples been processed.
The created dataframe will be merged with the spectral data during the fitting.
}
\keyword{correction}
\keyword{flattening}
\keyword{normalization}
