% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncHall300}
\alias{ncHall300}
\title{Hall sub data set from Hall et al. (2020)}
\format{
A nested list with 17 sub-lists:
\describe{
  \item{1 to 17}{A list containing the soil NMR spectrum of one of the following sites.
    \describe{
      \item{name}{"Calhoun", "CPER", "DCFS", "elve", "GRSM", "HARV", "icac", "JERC", "KONZ", "LENO", "MOAB", "NIWO", "ONAQ", "samt", "SCBI", "UNDE", "WOOD"}
      \item{NC}{Numeric vector.
      }
    }
  }
}
}
\source{
https://portal.edirepository.org/nis/mapbrowse?packageid=edi.575.1
}
\usage{
ncHall300
}
\description{
Contains 17 measurements of N and C, presented as molar N:C ratios.
}
\examples{
data(ncHall300)
}
\keyword{datasets}
