% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_data.R
\docType{data}
\name{nmr_data}
\alias{nmr_data}
\title{13C NMR data from}
\format{
A nested list containing all spectra from
}
\usage{
nmr_data
}
\description{
Reference
}
\examples{
data(spec)
}
\keyword{datasets}
