% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_table.R
\name{NMR_table}
\alias{NMR_table}
\title{Create a data frame of standard NMR areas}
\usage{
NMR_table(NMRmeth = NULL)
}
\arguments{
\item{ecosys}{Standards from the ecosystem to be fitted. "Terr_Nelson" or "Terr_Baldock" for terrestrial, "Aqua_Nelson" or "Aqua_Baldock" for aquatic.}
}
\description{
The function creates a data frame with all standards of the selected ecosystem (Terrestrial or Aquatic).
}
\keyword{standards}
