% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spec.R
\name{read_raw_spec}
\alias{read_raw_spec}
\title{Read spectra}
\usage{
read_raw_spec(file = NULL, filetype = NULL)
}
\arguments{
\item{file}{The raw  file}

\item{filetype}{The raw  file type "Bruker", .csv ("tab"), csv ("coma")}
}
\description{
This function reads the raw file, Bruker, tab separated or coma separated
extracts the spectra and returns a list with name, the raw spectrum
}
\examples{
## any .txt file as output from BRUKER
}
\keyword{integration}
