% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_nmr.R
\name{std_nmr}
\alias{std_nmr}
\title{Create a data frame of standard NMR areas}
\usage{
std_nmr(ecosys = NULL)
}
\arguments{
\item{ecosys}{Standards from the ecosystem to be fitted. "Terr_Nelson" or "Terr_Baldock" for terrestrial, "Aqua_Nelson" or "Aqua_Baldock" for aquatic.}
}
\description{
The function creates a data frame with all standards of the selected ecosystem (Terrestrial or Aquatic).
}
\examples{
std_table <- std_nmr(ecosys="Terr_Nelson")
}
\keyword{standards}
