% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_nc_data.R
\name{mk_nc_data}
\alias{mk_nc_data}
\title{Create .csv file for CN data}
\usage{
mk_nc_data(raw.spec)
}
\arguments{
\item{raw.spec}{The uploaded spectra read using the read_spec function}
}
\value{
A data frame with three columns, one containing the names extracted from the raw.spec, and two columns to be filled manually with the carbon and nitrogen values.
}
\description{
This function copies the spectra read using the read_spec function and creates
a .csv file with a colum with their names and two empty columns where the user must add the C and N values.
Thereafter the file is read with the function nc_data
}
\examples{
## any .txt file as output from BRUKER
}
\keyword{CN}
\keyword{file}
