% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_table.R
\name{NMR_table}
\alias{NMR_table}
\title{Create a data frame of standard NMR areas}
\usage{
NMR_table(NMRmeth = NULL)
}
\arguments{
\item{NMRmeth}{Regions to be integrated, methods available include: "4region", "Bonanomi", "Smernik" and Molecular mixing model ("MMM").}
}
\value{
A data frame containing the starting (from) and ending (to) ppm integral ranges of the different C functional groups.
}
\description{
The function creates a data frame with all standards of the selected ecosystem (Terrestrial or Aquatic).
}
\examples{
see_NMR_table <- NMR_table(NMRmeth="4region")
}
\keyword{standards}
