% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_ofset.R
\name{ssb_ofset}
\alias{ssb_ofset}
\title{Spinnning side bands ofset calculation function}
\usage{
ssb_ofset(NMRmeth = NULL, NMR_field = NULL, NMR_rotation = NULL)
}
\arguments{
\item{NMRmeth}{Regions to be integrated.
Default is spinning side bands, other methods available include: Bonanomi ("Bonanomi") and Molecular mixing model ("MMM" or "MMM-SSB").}

\item{raw.spec}{Raw spectrum}
}
\description{
This function calculates the spinning side band ofset for a given 13C NMR table.
The function returns the 13C NMR integration table to be used in the int_nmr function.
}
\keyword{integration}
\keyword{normalization,}
